refactor: 디렉토리 이름 변경
- user.ts 에서 backend.ts 로 변경
- backendApi 에서 api 디렉토리로 변경
