+++ src/components/chartSwiper/ChartSwiper.tsx
@@ -2,7 +2,7 @@ import clsx from 'clsx';
 import styles from './ChartSwiper.module.scss';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import { useQueries } from '@tanstack/react-query';
-import { getChartData, getIndicator } from '@/backendApi/fred';
+import { getChartData, getIndicator } from '@/api/fred';
 import const_queryKey from '@/const/queryKey';
 import LineChart from '../charts/line/LineChart';
 import { ChartDataForSwiperType, SeriessType, ValueType } from '@/types/fredType';
