+++ src/components/myContext/MyContext.tsx
@@ -3,7 +3,7 @@ import styles from './MyContext.module.scss';
 import { useSelector } from 'react-redux';
 import { Store } from '@/types/reduxType';
 import const_queryKey from '@/const/queryKey';
-import { getAllContexts, getContext, getContextIdsWithNames } from '@/backendApi/user';
+import { getAllContexts, getContext, getContextIdsWithNames } from '@/api/backend';
 import { useQueries, useQuery } from '@tanstack/react-query';
 import { ContextIdWithName, Journal, ContextType } from '@/types/userType';
 import CategoryTab from '../categoryTab/CategoryTab';
@@ -11,7 +11,7 @@ import { categoryNames } from '@/pages/_app';
 import IndicatorCard from '../cards/indicatorCard/IndicatorCard';
 import LineChart from '../charts/line/LineChart';
 import { useEffect, useState } from 'react';
-import { getChartData, getIndicator } from '@/backendApi/fred';
+import { getChartData, getIndicator } from '@/api/fred';
 import { Indicator } from '@/types/userType';
 import { cleanString } from '@/utils/cleanString';
 import { Swiper, SwiperSlide } from 'swiper/react';
